plugin update

+++ wp-content/plugins/gravityforms/entry_detail.php
@@ -1,4 +1,9 @@
 <?php
+
+if(!class_exists('GFForms')){
+    die();
+}
+
 class GFEntryDetail{
 
     public static function lead_detail_page(){
@@ -46,9 +51,14 @@ class GFEntryDetail{
                 $key       = $key_array[0];
                 $val       = $key_array[1] . ":" . $val;
             }
+            $type = rgget("type");
+            if(empty($type)){
+                $type = rgget("field_id") == "0" ? "global" : "field";
+            }
+
             $search_criteria["field_filters"][] = array(
                 "key" => $key,
-                "type" => rgempty("type", $_GET) ? "field" : rgget("type"),
+                "type" => $type,
                 "operator" => rgempty("operator", $_GET) ? "is" : rgget("operator"),
                 "value" => $val
             );
@@ -113,14 +123,15 @@ class GFEntryDetail{
                 RGFormsModel::add_note($lead["id"], $current_user->ID, $user_data->display_name, stripslashes($_POST["new_note"]));
 
                 //emailing notes if configured
-                if(rgpost("gentry_email_notes_to"))
-                {
+                if(rgpost("gentry_email_notes_to")){
+                	GFCommon::log_debug( 'Preparing to email entry notes.' );
                     $email_to = $_POST["gentry_email_notes_to"];
                     $email_from = $current_user->user_email;
                     $email_subject = stripslashes($_POST["gentry_email_subject"]);
-
                     $headers = "From: \"$email_from\" <$email_from> \r\n";
-                    $result = wp_mail($email_to, $email_subject, stripslashes($_POST["new_note"]), $headers);
+                    $body = stripslashes( $_POST["new_note"] );
+                    GFCommon::log_debug( "Emailing notes - TO: $email_to SUBJECT: $email_subject BODY: $body HEADERS: $headers" );
+                    $result = wp_mail( $email_to, $email_subject, $body, $headers );
                 }
             break;
 
@@ -157,6 +168,8 @@ class GFEntryDetail{
 
             case "delete" :
                 check_admin_referer('gforms_save_entry', 'gforms_save_entry');
+                if(!GFCommon::current_user_can_any("gravityforms_delete_entries"))
+                    die(__("You don't have adequate permissions to delete entries.", "gravityforms"));
                 RGFormsModel::delete_lead($lead["id"]);
                 ?>
                 <script type="text/javascript">
@@ -321,14 +334,14 @@ class GFEntryDetail{
                 <div id="side-info-column" class="inner-sidebar">
                 	<?php do_action("gform_entry_detail_sidebar_before", $form, $lead); ?>
 
+                    <!-- INFO BOX -->
                     <div id="submitdiv" class="stuffbox">
                         <h3>
-                            <span class="hndle"><?php _e("Info", "gravityforms"); ?></span>
+                            <span class="hndle"><?php _e("Entry", "gravityforms"); ?></span>
                         </h3>
                         <div class="inside">
                             <div id="submitcomment" class="submitbox">
                                 <div id="minor-publishing" style="padding:10px;">
-                                    <br/>
                                     <?php _e("Entry Id", "gravityforms"); ?>: <?php echo absint($lead["id"]) ?><br/><br/>
                                     <?php _e("Submitted on", "gravityforms"); ?>: <?php echo esc_html(GFCommon::format_date($lead["date_created"], false, "Y/m/d")) ?>
                                     <br/><br/>
@@ -354,37 +367,41 @@ class GFEntryDetail{
                                         <?php
                                     }
 
-                                    if(!empty($lead["payment_status"])){
-                                        echo $lead["transaction_type"] != 2 ? __("Payment Status", "gravityforms") : __("Subscription Status", "gravityforms"); ?>: <span id="gform_payment_status"><?php echo apply_filters("gform_payment_status", $lead["payment_status"], $form, $lead) ?></span>
-                                        <br/><br/>
-                                        <?php
-                                        if(!empty($lead["payment_date"])){
-                                            echo $lead["transaction_type"] == 1 ? __("Payment Date", "gravityforms") : __("Start Date", "gravityforms") ?>: <?php echo GFCommon::format_date($lead["payment_date"], false, "Y/m/d", $lead["transaction_type"] == 1) ?>
-                                            <br/><br/>
-                                            <?php
-                                        }
+                                    if(apply_filters("gform_enable_entry_info_payment_details", true, $lead)){
 
-                                        if(!empty($lead["transaction_id"])){
-                                            echo $lead["transaction_type"] == 1 ? __("Transaction Id", "gravityforms") : __("Subscriber Id", "gravityforms"); ?>: <?php echo $lead["transaction_id"]?>
+                                        if(!empty($lead["payment_status"])){
+                                            echo $lead["transaction_type"] == 2 ? __("Subscription Status", "gravityforms") : __("Payment Status", "gravityforms"); ?>: <span id="gform_payment_status"><?php echo apply_filters("gform_payment_status", $lead["payment_status"], $form, $lead) ?></span>
                                             <br/><br/>
                                             <?php
-                                        }
+                                            if(!empty($lead["payment_date"])){
+                                                echo $lead["transaction_type"] == 2 ? __("Start Date", "gravityforms") : __("Payment Date", "gravityforms") ?>: <?php echo GFCommon::format_date($lead["payment_date"], false, "Y/m/d", $lead["transaction_type"] != 2) ?>
+                                                <br/><br/>
+                                                <?php
+                                            }
 
-                                        if(!rgblank($lead["payment_amount"])){
-                                            echo $lead["transaction_type"] == 1 ? __("Payment Amount", "gravityforms") : __("Subscription Amount", "gravityforms"); ?>: <?php echo GFCommon::to_money($lead["payment_amount"], $lead["currency"]) ?>
-                                            <br/><br/>
-                                            <?php
+                                            if(!empty($lead["transaction_id"])){
+                                                echo $lead["transaction_type"] == 2 ? __("Subscriber Id", "gravityforms") : __("Transaction Id", "gravityforms"); ?>: <?php echo $lead["transaction_id"]?>
+                                                <br/><br/>
+                                                <?php
+                                            }
+
+                                            if(!rgblank($lead["payment_amount"])){
+                                                echo $lead["transaction_type"] == 2 ? __("Subscription Amount", "gravityforms") : __("Payment Amount", "gravityforms"); ?>: <?php echo GFCommon::to_money($lead["payment_amount"], $lead["currency"]) ?>
+                                                <br/><br/>
+                                                <?php
+                                            }
                                         }
                                     }
                                     do_action("gform_entry_info", $form["id"], $lead);
+
                                     ?>
                                 </div>
                                 <div id="major-publishing-actions">
-                                    <div>
+                                    <div id="delete-action">
                                         <?php
                                         switch($lead["status"]){
                                             case "spam" :
-                                                if(GFCommon::akismet_enabled($form['id'])){
+                                                if(GFCommon::spam_enabled($form['id'])){
                                                     ?>
                                                     <a onclick="jQuery('#action').val('unspam'); jQuery('#entry_form').submit()" href="#"><?php _e("Not Spam", "gravityforms") ?></a>
                                                     <?php
@@ -416,20 +433,15 @@ class GFEntryDetail{
                                                     ?>
                                                     <a class="submitdelete deletion" onclick="jQuery('#action').val('trash'); jQuery('#entry_form').submit()" href="#"><?php _e("Move to Trash", "gravityforms") ?></a>
                                                     <?php
-                                                    echo GFCommon::akismet_enabled($form['id']) ? "|" : "";
+                                                    echo GFCommon::spam_enabled($form['id']) ? "|" : "";
                                                 }
-                                                if(GFCommon::akismet_enabled($form['id'])){
+                                                if(GFCommon::spam_enabled($form['id'])){
                                                 ?>
                                                     <a class="submitdelete deletion" onclick="jQuery('#action').val('spam'); jQuery('#entry_form').submit()" href="#"><?php _e("Mark as Spam", "gravityforms") ?></a>
                                                 <?php
                                                 }
 
                                         }
-
-                                        /*if(GFCommon::current_user_can_any("gravityforms_delete_entries")){
-                                            $delete_link = '<a class="submitdelete deletion" onclick="if ( confirm(\''. __("You are about to delete this entry. \'Cancel\' to stop, \'OK\' to delete.", "gravityforms") .'\') ) { jQuery(\'#action\').val(\'delete\'); jQuery(\'#entry_form\')[0].submit();} return false;" href="#">' . __("Delete", "gravityforms") . '</a>';
-                                            echo apply_filters("gform_entrydetail_delete_link", $delete_link);
-                                        }*/
 
                                         ?>
                                     </div>
@@ -445,10 +457,18 @@ class GFEntryDetail{
                                             }
                                         ?>
                                     </div>
+                                    <div class="clear"></div>
                                 </div>
                             </div>
                         </div>
                     </div>
+
+                    <?php
+                    if (!empty($lead["payment_status"]) && !apply_filters("gform_enable_entry_info_payment_details", true, $lead) ) {
+                        self::payment_details_box($lead, $form);
+                    }
+                    ?>
+
                     <?php do_action("gform_entry_detail_sidebar_middle", $form, $lead); ?>
 
                     <?php if(GFCommon::current_user_can_any("gravityforms_edit_entry_notes") ) { ?>
@@ -526,7 +546,7 @@ class GFEntryDetail{
                         do_action("gform_entry_detail", $form, $lead);
 
                         if(GFCommon::current_user_can_any("gravityforms_view_entry_notes")) { ?>
-                            <div id="namediv" class="stuffbox">
+                            <div class="postbox">
                                 <h3>
                                     <label for="name"><?php _e("Notes", "gravityforms"); ?></label>
                                 </h3>
@@ -547,8 +567,8 @@ class GFEntryDetail{
                                             }
                                         }
                                         //displaying notes grid
-                                        $subject = !empty($form["autoResponder"]["subject"]) ? "RE: " . GFCommon::replace_variables($form["autoResponder"]["subject"], $form, $lead) : "";
-                                        self::notes_grid($notes, true, $emails, $subject);
+                                        $subject = '';
+                                        self::notes_grid( $notes, true, $emails, $subject );
                                         ?>
                                     </div>
                                 </form>
@@ -576,7 +596,7 @@ class GFEntryDetail{
         $form = apply_filters("gform_admin_pre_render_" . $form["id"], apply_filters("gform_admin_pre_render", $form));
         $form_id = $form["id"];
         ?>
-        <div id="namediv" class="stuffbox">
+        <div class="postbox">
             <h3>
                 <label for="name"><?php _e("Details", "gravityforms"); ?></label>
             </h3>
@@ -629,7 +649,7 @@ class GFEntryDetail{
         <?php
     }
 
-    public static function notes_grid($notes, $is_editable, $emails = null, $autoresponder_subject=""){
+    public static function notes_grid( $notes, $is_editable, $emails = null, $subject = '' ){
         if(sizeof($notes) > 0 && $is_editable && GFCommon::current_user_can_any("gravityforms_edit_entry_notes")){
             ?>
             <div class="alignleft actions" style="padding:3px 0;">
@@ -681,14 +701,15 @@ class GFEntryDetail{
                         <td class="entry-detail-note<?php echo $is_last ? " lastrow" : "" ?>">
                     <?php
                     }
+                    $class = $note->note_type ? " gforms_note_{$note->note_type}" : "";
                     ?>
                             <div style="margin-top:4px;">
-                                <div class="note-avatar"><?php echo get_avatar($note->user_id, 48);?></div>
+                                <div class="note-avatar"><?php echo apply_filters("gform_notes_avatar", get_avatar($note->user_id, 48), $note);?></div>
                                 <h6 class="note-author"> <?php echo esc_html($note->user_name)?></h6>
                                 <p style="line-height:130%; text-align:left; margin-top:3px;"><a href="mailto:<?php echo esc_attr($note->user_email)?>"><?php echo esc_html($note->user_email) ?></a><br />
                                 <?php _e("added on", "gravityforms"); ?> <?php echo esc_html(GFCommon::format_date($note->date_created, false)) ?></p>
                             </div>
-                            <div class="detail-note-content"><?php echo esc_html($note->value) ?></div>
+                            <div class="detail-note-content<?php echo $class ?>"><?php echo esc_html($note->value) ?></div>
                         </td>
 
                 </tr>
@@ -716,7 +737,7 @@ class GFEntryDetail{
 
                                 <span id='gentry_email_subject_container' style="display:none;">
                                     <label for="gentry_email_subject"><?php _e("Subject:", "gravityforms") ?></label>
-                                    <input type="text" name="gentry_email_subject" id="gentry_email_subject" value="<?php echo $autoresponder_subject ?>" style="width:35%"/>
+									<input type="text" id="gentry_email_subject" name="gentry_email_subject" value="" style="width:35%">
                                 </span>
                             </span>
                         <?php } ?>
@@ -830,10 +851,10 @@ class GFEntryDetail{
                             <td colspan="2" class="entry-view-field-value lastrow">
                                 <table class="entry-products" cellspacing="0" width="97%">
                                     <colgroup>
-                                          <col class="entry-products-col1">
-                                          <col class="entry-products-col2">
-                                          <col class="entry-products-col3">
-                                          <col class="entry-products-col4">
+                                          <col class="entry-products-col1" />
+                                          <col class="entry-products-col2" />
+                                          <col class="entry-products-col3" />
+                                          <col class="entry-products-col4" />
                                     </colgroup>
                                     <thead>
                                         <th scope="col"><?php echo apply_filters("gform_product_{$form_id}", apply_filters("gform_product", __("Product", "gravityforms"), $form_id), $form_id) ?></th>
@@ -924,6 +945,74 @@ class GFEntryDetail{
 
         return '<a ' . $href . ' class="' . $class . '" title="' . $label . '"><i class="fa-lg ' . $icon . '"></i></a></li>';
     }
+
+    /**
+     * @param $lead
+     * @param $form
+     * @return mixed
+     */
+    public static function payment_details_box($lead, $form)
+    {
+        ?>
+        <!-- PAYMENT BOX -->
+        <div id="submitdiv" class="stuffbox">
+            <h3>
+                <span
+                    class="hndle"><?php echo $lead["transaction_type"] == 2 ? __("Subscription Details", "gravityforms") : __("Payment Details", "gravityforms"); ?></span>
+            </h3>
+
+            <div class="inside">
+                <div id="submitcomment" class="submitbox">
+                    <div id="minor-publishing" style="padding:10px;">
+                        <?php
+							if (!empty($lead["payment_status"])) {
+							?>
+								<div id="gf_payment_status" class="gf_payment_detail">
+								<?php
+									echo __("Status", "gravityforms"); ?>: <span id="gform_payment_status"><?php echo apply_filters("gform_payment_status", $lead["payment_status"], $form, $lead)
+								?></span>
+								</div>
+
+							<?php
+                            if (!empty($lead["payment_date"])) {
+								?>
+								<div id="gf_payment_date" class="gf_payment_detail">
+									<?php
+                                	echo $lead["transaction_type"] == 2 ? __("Start Date", "gravityforms") : __("Date", "gravityforms") ?>: <?php echo GFCommon::format_date($lead["payment_date"], false, "Y/m/d", $lead["transaction_type"] != 2)
+                                	?>
+								</div>
+                            <?php
+                            }
+
+                            if (!empty($lead["transaction_id"])) {
+								?>
+								<div id="gf_payment_transaction_id" class="gf_payment_detail">
+								<?php
+                                echo $lead["transaction_type"] == 2 ? __("Subscription Id", "gravityforms") : __("Transaction Id", "gravityforms"); ?>: <?php echo $lead["transaction_id"] ?>
+                                </div>
+                            <?php
+                            }
+
+                            if (!rgblank($lead["payment_amount"])) {
+								?>
+								<div id="gf_payment_amount" class="gf_payment_detail">
+									<?php
+                                echo $lead["transaction_type"] == 2 ? __("Recurring Amount", "gravityforms") : __("Amount", "gravityforms"); ?>: <?php echo GFCommon::to_money($lead["payment_amount"], $lead["currency"]) ?>
+                                </div>
+                            <?php
+                            }
+                        }
+                        do_action("gform_payment_details", $form["id"], $lead);
+
+                        ?>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <?php
+    }
+
+
 
 }
 ?>
